Restructure http requests, Add DressSale model and view

+++ cordelia/admin.py
@@ -2,8 +2,8 @@ from flask import Blueprint, render_template, redirect, url_for, flash, request
 from sqlalchemy import case, desc, func
 from cordelia.auth import login_required, admin_required
 from cordelia.db import db
-from cordelia.models import Dress, Customer, Rent, Maintenance, maintenance_association
-from cordelia.forms import SearchForm, DressForm, RentForm, CustomerForm, MaintenanceForm, DeleteForm
+from cordelia.models import Dress, Customer, Rent, Maintenance, maintenance_association, DressSale
+from cordelia.forms import SearchForm, DressForm, RentForm, CustomerForm, MaintenanceForm, DeleteForm, SaleForm
 from base64 import b64encode
 
 
@@ -13,7 +13,7 @@ import logging
 adminBp = Blueprint('admin', __name__, url_prefix='/admin')
 
 
-@adminBp.route('/download/excel')
+@adminBp.route('/download/db-xlsx')
 @login_required
 @admin_required
 def downloadExcel():
@@ -21,26 +21,6 @@ def downloadExcel():
     from cordelia.excel import excel_download
 
     return excel_download()
-
-
-@adminBp.route('/display_plot/<int:img_num>', methods=['GET', 'POST'])
-@login_required
-@admin_required
-def display_plot(img_num):
-    if Dress.query.first() and Rent.query.first() and Customer.query.first() and Maintenance.query.first():
-        
-        from cordelia.dataframe import costs_vs_earnings, top_customers, plot_combined_statistics
-
-        if img_num == 1:
-            image_data = costs_vs_earnings()  
-        elif img_num == 2:
-            image_data = top_customers()
-        elif img_num == 3:
-            image_data = plot_combined_statistics()
-    
-        return render_template('admin_views/plots_template.html', image_data=image_data)
-    else:
-        return render_template('admin_views/plots_template.html')
 
 
 # SearchForm handler
@@ -73,13 +53,13 @@ def update_statuses_endpoint():
         flash('Statuses updated successfully.', 'success')
 
     logging.debug("update_statuses() method called from dashboard")
-    return redirect(url_for('admin.inventory'))
+    return redirect(url_for('admin.dress_db'))
 
 
 
-@adminBp.route('/inventory', methods=['GET', 'POST'])
+@adminBp.route('/dashboard/dress-db', methods=['GET', 'POST'])
 @login_required
-def inventory():
+def dress_db():
     model = 'Dress'
     
     model_columns = Dress.__table__.columns.keys()
@@ -89,17 +69,17 @@ def inventory():
     items_per_page = 12
 
     # Get the selected column for ordering from the query parameters
-    order_by_column = request.args.get('order_by', default='default')
+    order_by_column = request.args.get('sort', default='status')
 
     # Initial inventory query
     inventory_query = Dress.query
 
     # Apply sorting based on the selected column
-    if order_by_column == 'times_rented':
+    if order_by_column == 'popularity':
         inventory_query = inventory_query.order_by(Dress.timesRented.desc())
     elif order_by_column == 'id':
         inventory_query = inventory_query.order_by(Dress.id)
-    elif order_by_column == 'dress_cost':
+    elif order_by_column == 'cost':
         inventory_query = inventory_query.order_by(Dress.cost.desc())
     else:
         # If 'default', sort by rentStatus and maintenanceStatus in descending order.
@@ -143,6 +123,8 @@ def inventory():
 
     maintenance_form = MaintenanceForm()
 
+    sell_dress_form = SaleForm()
+
     # Handle form submissions for the delete form
     if delete_form.validate_on_submit():
         dress_id = delete_form.id.data
@@ -151,21 +133,27 @@ def inventory():
     # Handle form submissions for adding dresses to maintenance_form
     if maintenance_form.validate_on_submit():
         return redirect(url_for('admin.add_maintenance'))
+    
+    # Handle form submissions for dress sales form
+    if sell_dress_form.validate_on_submit():
+        dress_id = sell_dress_form.id.data
+        return redirect(url_for('admin.sell_dress', id=dress_id))
 
-    return render_template('admin_views/inventory_dress.html',
+    return render_template('admin_views/db_dress.html',
                            inventory=inventory,
                            form=form,
                            maintenance_form=maintenance_form,
                            delete_form=delete_form,
+                           sell_dress_form=sell_dress_form,
                            pagination=pagination,
                            order_by_column=order_by_column,
                            model=model)
 
 
 
-@adminBp.route('/maintenances', methods=['GET','POST'])
+@adminBp.route('/dashboard/maintenance-db', methods=['GET', 'POST'])
 @login_required
-def maintenance_inventory():
+def maintenance_db():
     model = 'Maintenance'
 
     model_columns = Maintenance.__table__.columns.keys()
@@ -173,7 +161,7 @@ def maintenance_inventory():
     page = request.args.get('page', 1, type=int)
     items_per_page = 12
 
-    order_by_column = request.args.get('order_by', default='default')
+    order_by_column = request.args.get('sort', default='status')
 
     inventory_query = Maintenance.query
 
@@ -198,7 +186,7 @@ def maintenance_inventory():
         maintenance_id = delete_form.id.data
         return redirect(url_for('admin.delete_object', dataBase='Maintenance', id=maintenance_id))
 
-    return render_template('admin_views/inventory_maintenance.html', 
+    return render_template('admin_views/db_maintenance.html', 
                            inventory=inventory, 
                            form=form, 
                            delete_form=delete_form, 
@@ -208,9 +196,9 @@ def maintenance_inventory():
 
 
 
-@adminBp.route('/rent-inventory', methods=['GET', 'POST'])
+@adminBp.route('/dashboard/rent-db', methods=['GET', 'POST'])
 @login_required
-def rentInventory():
+def rent_db():
     model = 'Rent'
 
     model_columns = Rent.__table__.columns.keys()
@@ -218,13 +206,13 @@ def rentInventory():
     page = request.args.get('page', 1, type=int)
     items_per_page = 12
 
-    order_by_column = request.args.get('order_by', default='default')
+    order_by_column = request.args.get('sort', default='status')
 
     inventory_query = Rent.query
 
     if order_by_column == 'id':
         inventory_query = inventory_query.order_by(Rent.id)
-    elif order_by_column == 'user_lastName':
+    elif order_by_column == 'customer_last_name':
         # Join Rent and Customer tables and sort by Customer's lastName
         inventory_query = inventory_query.join(Customer, Rent.clientId == Customer.id).order_by(Customer.lastName)
     elif order_by_column == 'dress_id':
@@ -244,7 +232,7 @@ def rentInventory():
         rent_id = delete_form.id.data
         return redirect(url_for('admin.delete_object', dataBase='Rent', id=rent_id))
 
-    return render_template('admin_views/inventory_rent.html', 
+    return render_template('admin_views/db_rent.html', 
                            inventory=inventory, 
                            form=form, 
                            delete_form=delete_form, 
@@ -254,9 +242,9 @@ def rentInventory():
 
 
 
-@adminBp.route('/customer-inventory', methods=["GET", "POST"])
+@adminBp.route('/dashboard/customer-db', methods=["GET", "POST"])
 @login_required
-def customerInventory():
+def customer_db():
     model = 'Customer'
 
     model_columns = Customer.__table__.columns.keys()
@@ -264,7 +252,7 @@ def customerInventory():
     page = request.args.get('page', 1, type=int)
     items_per_page = 12
 
-    order_by_column = request.args.get('order_by', default='default')
+    order_by_column = request.args.get('sort', default='status')
 
     inventory_query = Customer.query
 
@@ -304,7 +292,7 @@ def customerInventory():
         customer_id = delete_form.id.data
         return redirect(url_for('admin.delete_object', dataBase='Customer', id=customer_id))
 
-    return render_template('admin_views/inventory_customer.html', 
+    return render_template('admin_views/db_customer.html', 
                            inventory=inventory, 
                            form=form, 
                            delete_form=delete_form, 
@@ -314,7 +302,52 @@ def customerInventory():
 
 
 
-@adminBp.route("/update-db/<string:title>/<string:form_type>", methods=['GET', 'POST'])
+@adminBp.route('/dashboard/sales-db', methods=['GET', 'POST'])
+@login_required
+def sales_db():
+    model = 'DressSale'
+
+    model_columns = DressSale.__table__.columns.keys()
+
+    page = request.args.get('page', 1, type=int)
+    items_per_page = 12
+
+    order_by_column = request.args.get('sort', default='id')
+
+    inventory_query = DressSale.query
+
+    if order_by_column =='date':
+        inventory_query = inventory_query.order_by(DressSale.sale_date.desc())
+    elif order_by_column == 'customer':
+        inventory_query = inventory_query.order_by(DressSale.customer_id)
+    elif order_by_column == 'price':
+        inventory_query = inventory_query.order_by(DressSale.sale_price.desc())
+    else:
+        inventory_query = inventory_query.order_by(DressSale.id.desc())
+
+    inventory_query, form = handle_search_form(inventory_query, model_columns, DressSale)
+
+    inventory = inventory_query.paginate(page=page, per_page=items_per_page)
+
+    pagination = inventory_query.paginate(page=page, per_page=items_per_page)
+
+    delete_form = DeleteForm()
+
+    if delete_form.validate_on_submit():
+        sale_id = delete_form.id.data
+        return redirect(url_for('admin.delete_object', dataBase='DressSale', id=sale_id))
+
+    return render_template('admin_views/db_sales.html', 
+                           inventory=inventory, 
+                           form=form, 
+                           delete_form=delete_form, 
+                           pagination=pagination, 
+                           order_by_column=order_by_column, 
+                           model=model)
+
+
+
+@adminBp.route("/dashboard/update-db/<string:title>/<string:form_type>", methods=['GET','POST'])
 @login_required
 def update(title, form_type):
     # Initialize form based on form_type
@@ -326,7 +359,7 @@ def update(title, form_type):
         form = RentForm()
     else:
         flash('Invalid form type.', 'danger')
-        return redirect(url_for('admin.inventory'))
+        return redirect(url_for('home.html'))
 
     if form.validate_on_submit():
 
@@ -344,7 +377,7 @@ def update(title, form_type):
             flash(f'{customer} added successfully into the database.')
             logging.debug(f"{customer} Added from dashboard")
             
-            return redirect(url_for('admin.customerInventory'))
+            return redirect(url_for('admin.customer_db'))
 
         if form_type == 'dress':
             # Get the image data from the request
@@ -375,7 +408,7 @@ def update(title, form_type):
             flash(f'{dress} added successfully into the database.')
             logging.debug(f"{dress} Added from dashboard")
             
-            return redirect(url_for('admin.inventory'))
+            return redirect(url_for('admin.dress_db'))
         
         elif form_type == 'rent':
 
@@ -385,7 +418,7 @@ def update(title, form_type):
             dress = Dress.query.get(dress_id)
             customer = Customer.query.get(customer_id)
 
-            if not dress.check_status() and not dress.check_maintenance_status():
+            if not dress.check_status() and not dress.check_maintenance_status() and not dress.sold:
                 rent = Rent(
                     dressId=form.dressId.data,
                     clientId=form.customerId.data,
@@ -410,7 +443,7 @@ def update(title, form_type):
                 flash(f'{rent} added successfully into the database.')
                 logging.debug(f"{rent} Added from dashboard")
 
-                return redirect(url_for('admin.rentInventory')) 
+                return redirect(url_for('admin.rent_db')) 
             else:
                 flash('Dress not available.')
 
@@ -418,7 +451,7 @@ def update(title, form_type):
 
 
 
-@adminBp.route('/set-maintenance', methods=['POST'])
+@adminBp.route('/add-maintenance', methods=['POST'])
 @login_required
 @admin_required
 def add_maintenance():
@@ -429,11 +462,11 @@ def add_maintenance():
         dress_id = dress_id_form.dress_id.data
         if dress_id:
             dress = Dress.query.get(dress_id)
-            if dress and not dress.check_status() and not dress.check_maintenance_status():
+            if dress and not dress.check_status() and not dress.check_maintenance_status() and not dress.sold:
                 selected_dresses.append(dress)
             else:
                 flash('One or more dresses are unavaiable.', 'error')
-                return redirect(url_for('admin.inventory'))
+                return redirect(url_for('admin.dress_db'))
 
     date = form.maintenanceDate.data
     type = form.maintenanceType.data
@@ -455,7 +488,7 @@ def add_maintenance():
     db.session.commit()
     flash('Maintenance record updated successfully!', 'success')
         
-    return redirect(url_for('admin.inventory'))
+    return redirect(url_for('admin.dress_db'))
 
 
 
@@ -467,7 +500,7 @@ def delete_object(dataBase, id):
     if dataBase == 'Dress':
         dress = Dress.query.get(int(id))
 
-        if dress and not dress.check_status() and not dress.check_maintenance_status():
+        if dress and not dress.check_status() and not dress.check_maintenance_status() and not dress.sold:
             db.session.delete(dress)
             db.session.commit()
             flash(f'{dress} deleted successfully.', 'success')
@@ -475,7 +508,7 @@ def delete_object(dataBase, id):
         else:
             flash('Dress not available.', 'danger')
     
-        return redirect(url_for('admin.inventory'))
+        return redirect(url_for('admin.dress_db'))
     
     elif dataBase == 'Rent':
         rent = Rent.query.get(int(id))
@@ -502,7 +535,7 @@ def delete_object(dataBase, id):
             flash(f'{rent} deleted successfully.', 'success')
             logging.debug(f"{rent} deleted from dashboard")
 
-            return redirect(url_for('admin.rentInventory'))
+            return redirect(url_for('admin.rent_db'))
         else:
             flash('Rent not found.', 'danger')
     
@@ -516,7 +549,7 @@ def delete_object(dataBase, id):
             flash(f'{customer} deleted successfully.', 'success')
             logging.debug(f"{customer} deleted from dashboard")
 
-            return redirect(url_for('admin.customerInventory'))
+            return redirect(url_for('admin.customer_db'))
         else:
             flash("Can't delete this customer.", 'danger')
 
@@ -539,6 +572,78 @@ def delete_object(dataBase, id):
             flash(f'{maintenance} deleted successfully.', 'success')
             logging.debug(f"{maintenance} deleted from dashboard")
 
-            return redirect(url_for('admin.maintenance_inventory'))
+            return redirect(url_for('admin.maintenance_db'))
         else:
             flash("Cannot delete this maintenance.", 'danger')
+
+    elif dataBase == 'DressSale':
+        sale = DressSale.query.get(int(id))
+
+        if sale:
+            db.session.delete(sale)
+
+            db.session.commit()
+
+            flash(f'{sale} deleted successfully.', 'success')
+            logging.debug(f"{sale} deleted from dashboard")
+
+            return redirect(url_for('admin.sales_db'))
+        else:
+            flash("Cannot delete this sale.", 'danger')
+
+
+
+
+@adminBp.route('/sell-dress/', methods=['GET', 'POST'])
+def sell_dress():
+    form = SaleForm()
+
+    if form.validate_on_submit():
+        dress_id = form.dress_id.data
+        dress = Dress.query.get_or_404(dress_id)
+
+        if not dress.sold and not dress.rentStatus and not dress.maintenanceStatus:
+            # Mark the dress as sold
+            dress.sold = True
+
+            # Create a new entry in the DressSale model
+            sale = DressSale(
+                dress_id=dress.id,
+                customer_id=form.customer_id.data,
+                sale_date=form.sale_date.data,
+                sale_price=form.sale_price.data
+            )
+
+            db.session.add(sale)
+            db.session.commit()
+
+            flash('Dress sold successfully', 'success')
+            return redirect(url_for('admin.dress_db'))
+        else:
+            flash('Dress is already sold or has other active status', 'warning')
+
+    return redirect(url_for('admin.dress_db'))
+
+    
+
+
+
+
+@adminBp.route('/dashboard/plot/<int:img_num>', methods=['GET'])
+@login_required
+@admin_required
+def display_plot(img_num):
+    if Dress.query.first() and Rent.query.first() and Customer.query.first() and Maintenance.query.first():
+        
+        from cordelia.plots import costs_vs_earnings, top_customers, plot_combined_statistics
+
+        if img_num == 1:
+            image_data = costs_vs_earnings()  
+        elif img_num == 2:
+            image_data = top_customers()
+        elif img_num == 3:
+            image_data = plot_combined_statistics()
+    
+        return render_template('admin_views/plots.html', image_data=image_data)
+    else:
+        return render_template('admin_views/plots.html')
